refact: rewrite horizontal form group via typescript
